PLZ READ. Fixed alt tags in roster.
- Fixed orginial en_roster.html

**IMPORTANT**

This meant I had to create a NodeJS Environement to run the script that automatically replaced the alt tags because I imported libraries to help out. Maybe it would've been faster to change manually but it's cooler like this. Also it opens up the possibility to compress out files (Next commit)

BTW it won't affect our website